requirements.txt
Django==2.0.7
django_payu==0.5
payu==0.5.1
djangorestframework==3.11.0
